Add unit tests for color conversion functions
